backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
tiktoken==0.7.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
